ref(parse): move payload type to parse module
