+++ src/derivation/attached_signature_code.rs
@@ -1,6 +1,6 @@
 use super::{self_signing::SelfSigning, DerivationCode};
-use crate::{error::Error, event_message::payload_size::num_to_b64};
-use base64::decode_config;
+use crate::error::Error;
+use base64::{decode_config, encode_config};
 use core::str::FromStr;
 
 /// Attached Signature Derivation Codes
@@ -91,4 +91,24 @@ pub fn b64_to_num(b64: &[u8]) -> Result<u16, Error> {
         1 => [0, slice[0]],
         _ => [slice[len - 2], slice[len - 1]],
     }))
+}
+
+pub fn num_to_b64(num: u16) -> String {
+    match num {
+        n if n < 63 => {
+            encode_config([num.to_be_bytes()[1] << 2], base64::URL_SAFE_NO_PAD)[..1].to_string()
+        }
+        n if n < 4095 => encode_config(num.to_be_bytes(), base64::URL_SAFE_NO_PAD)[..2].to_string(),
+        _ => encode_config(num.to_be_bytes(), base64::URL_SAFE_NO_PAD),
+    }
+}
+
+#[test]
+fn num_to_b64_test() {
+    assert_eq!("A", num_to_b64(0));
+    assert_eq!("B", num_to_b64(1));
+    assert_eq!("C", num_to_b64(2));
+    assert_eq!("D", num_to_b64(3));
+    assert_eq!("b", num_to_b64(27));
+    assert_eq!("AE", num_to_b64(64));
 }
